examples/anthropic-claude-code-in-sandbox-python/pyproject.toml
[project]
name = "anthropic-claude-code-in-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "e2b>=2.1.0",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
